split the logic between auth and userdata DAO
