+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/extensions/DaoExtension.java
@@ -12,8 +12,7 @@ public class DaoExtension implements TestInstancePostProcessor {
     @Override
     public void postProcessTestInstance(Object testInstance, ExtensionContext context) throws Exception {
         for (Field field : testInstance.getClass().getDeclaredFields()) {
-            if ((field.getType().isAssignableFrom(AuthUserDAO.class) || field.getType().isAssignableFrom(UserDataUserDAO.class))
-                    && field.isAnnotationPresent(Dao.class)) {
+            if (field.getType().isAssignableFrom(AuthUserDAO.class) && field.isAnnotationPresent(Dao.class)) {
                 field.setAccessible(true);
 
                 AuthUserDAO authUserDAO;
@@ -27,6 +26,22 @@ public class DaoExtension implements TestInstancePostProcessor {
                 }
 
                 field.set(testInstance, authUserDAO);
+            }
+
+            if (field.getType().isAssignableFrom(UserDataUserDAO.class) && field.isAnnotationPresent(Dao.class)) {
+                field.setAccessible(true);
+
+                UserDataUserDAO userDataUserDAO;
+
+                if ("hibernate".equals(System.getProperty("db.impl"))) {
+                    userDataUserDAO = new UserDataUserDAOJdbc();
+                } else if ("spring".equals(System.getProperty("db.impl"))) {
+                    userDataUserDAO = new UserDataUserDAOJdbc();
+                } else {
+                    userDataUserDAO = new UserDataUserDAOJdbc();
+                }
+
+                field.set(testInstance, userDataUserDAO);
             }
 
         }
